+++ include/spica.h
@@ -30,7 +30,7 @@
 
 #include "../src/renderer/pathtrace.h"
 #include "../src/renderer/bdpt.h"
-#include "../src/renderer/mlt_renderer.h"
+#include "../src/renderer/mlt.h"
 #include "../src/renderer/ppmprob.h"
 #include "../src/renderer/sppm.h"
 #include "../src/renderer/subsurface_sppm_renderer.h"
