+++ src/renderer/CMakeLists.txt
@@ -11,7 +11,7 @@ set(SOURCES
     ${CMAKE_CURRENT_LIST_DIR}/renderer_helper.cc
     ${CMAKE_CURRENT_LIST_DIR}/pathtrace.cc
     ${CMAKE_CURRENT_LIST_DIR}/bdpt.cc
-#    ${CMAKE_CURRENT_LIST_DIR}/mlt_renderer.cc
+    ${CMAKE_CURRENT_LIST_DIR}/mlt.cc
     ${CMAKE_CURRENT_LIST_DIR}/photon_map.cc
     ${CMAKE_CURRENT_LIST_DIR}/ppmprob.cc
     ${CMAKE_CURRENT_LIST_DIR}/sppm.cc
@@ -34,7 +34,7 @@ set(HEADERS
     ${CMAKE_CURRENT_LIST_DIR}/renderer_helper.h
     ${CMAKE_CURRENT_LIST_DIR}/pathtrace.h
     ${CMAKE_CURRENT_LIST_DIR}/bdpt.h
-#    ${CMAKE_CURRENT_LIST_DIR}/mlt_renderer.h
+    ${CMAKE_CURRENT_LIST_DIR}/mlt.h
     ${CMAKE_CURRENT_LIST_DIR}/photon_map.h
     ${CMAKE_CURRENT_LIST_DIR}/ppmprob.h
     ${CMAKE_CURRENT_LIST_DIR}/sppm.h
